Renamed about and landing page
